refactor: move useUser and useToken to hooks' folder in root
